backend/requirements.txt
# ── FastAPI 코어 ──
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-multipart>=0.0.9   # 파일 업로드 (UploadFile)
httpx>=0.27.0             # 비동기 HTTP 클라이언트
websockets>=12.0          # Binance WebSocket

# ── OpenAI (Vision, Whisper STT, TTS, Embedding) ──
openai>=1.35.0

# ── Google Gemini (Vision fallback) ──
google-genai>=1.0.0

# ── RAG: LangChain + ChromaDB ──
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-chroma>=0.1.0
chromadb>=0.5.0

# ── PPO 강화학습 ──
stable-baselines3>=2.3.0
gymnasium>=0.29.0
numpy>=1.26.0
torch>=2.3.0               # stable-baselines3 의존성

# ── 데이터 처리 ──
pandas>=2.2.0

# ── 스케줄러 (n8n 스케줄 트리거 대체) ──
apscheduler>=3.10.0

# ── 환경 변수 ──
python-dotenv>=1.0.0
